Replaced ApproxPipelineSpec with  ApproximationAssertions
